Hydroelastic potato contact
To run this:

  bazel run //examples/multibody/potato:hydro_potato

The peeler will penetrate the potato and a contat mesh will be drawn, but
the two objects will *not* separate. Because there are no contact *forces*.

To get contacts:

  bazel run //examples/multibody/potato:hydro_potato -- --rez_hint 0.019

(Or any number strictly less than 0.02). In three *visual* steps the peeler
and potato will separate.

The coarsest possible box has *no* internal vertices so that means the
pressure on the contact surface will always be zero, therefore no force.